Remove the webhook
The Github Actions version of this code doesn't need the hook and the
hook related code makes things a bit more complex to change. This also
removes the code that fetches the servo repository, as it is now the
responsibility of the test or the action to ensure an updated local
servo checkout.