Use the local servo checkout to get upstreamable commits
Now that we know that the local servo checkout will always have the
current commits, we don't need to fetch them from GitHub. This avoids
one HTTP request per commit and reduces the code used for testing.
Instead we simply add the commits to the mock repositories when tests
start.